Update NHMetadataBuilder.cs
In case of a mapped string collection the property ElementPersister of BasicCollectionPersister throws a NHibernate.AssertionFailure with text "Not an association". Therefore, I have skipped that case by the conditions IsManyToMany or IsOneToMany which both return false for string collections.  Attention: This code is probably better than the original in avoiding an exception, but I am not able to assess whether the skipped code has to substituted by something else.